// clients/admin-ui/src/types/api/models/DataPurposeResponse.ts
// This file is auto-generated by @hey-api/openapi-ts

/**
 * DataPurposeResponse
 */
export type DataPurposeResponse = {
  /**
   * Id
   */
  id: string;
  /**
   * Fides Key
   */
  fides_key: string;
  /**
   * Name
   */
  name: string;
  /**
   * Description
   */
  description?: string | null;
  /**
   * Organization Fides Key
   */
  organization_fides_key?: string | null;
  /**
   * Tags
   */
  tags?: Array<string> | null;
  /**
   * Data Use
   */
  data_use: string;
  /**
   * Data Subject
   */
  data_subject?: string | null;
  /**
   * Data Categories
   */
  data_categories?: Array<string>;
  /**
   * Legal Basis For Processing
   */
  legal_basis_for_processing?: string | null;
  /**
   * Flexible Legal Basis For Processing
   */
  flexible_legal_basis_for_processing?: boolean;
  /**
   * Special Category Legal Basis
   */
  special_category_legal_basis?: string | null;
  /**
   * Impact Assessment Location
   */
  impact_assessment_location?: string | null;
  /**
   * Retention Period
   */
  retention_period?: string | null;
  /**
   * Features
   */
  features?: Array<string>;
  /**
   * Created At
   */
  created_at: string;
  /**
   * Updated At
   */
  updated_at: string;
};